+++ packages/chakra-ui/src/Templates/Templates.ts
@@ -4,7 +4,7 @@ import ArrayFieldTemplate from '../ArrayFieldTemplate';
 import BaseInputTemplate from '../BaseInputTemplate/BaseInputTemplate';
 import DescriptionField from '../DescriptionField';
 import ErrorList from '../ErrorList';
-import { MoveDownButton, MoveUpButton, RemoveButton } from '../IconButton';
+import { CopyButton, MoveDownButton, MoveUpButton, RemoveButton } from '../IconButton';
 import FieldErrorTemplate from '../FieldErrorTemplate';
 import FieldHelpTemplate from '../FieldHelpTemplate';
 import FieldTemplate from '../FieldTemplate';
@@ -24,6 +24,7 @@ export function generateTemplates<
     ArrayFieldTemplate,
     BaseInputTemplate,
     ButtonTemplates: {
+      CopyButton,
       AddButton,
       MoveDownButton,
       MoveUpButton,