+++ packages/material-ui/src/IconButton/IconButton.tsx
@@ -1,6 +1,7 @@
 import IconButton, { IconButtonProps as MuiIconButtonProps } from '@material-ui/core/IconButton';
 import ArrowUpwardIcon from '@material-ui/icons/ArrowUpward';
 import ArrowDownwardIcon from '@material-ui/icons/ArrowDownward';
+import CopyIcon from '@material-ui/icons/FileCopy';
 import RemoveIcon from '@material-ui/icons/Remove';
 import { FormContextType, IconButtonProps, RJSFSchema, StrictRJSFSchema, TranslatableString } from '@rjsf/utils';
 
@@ -14,6 +15,21 @@ export default function MuiIconButton<
     <IconButton {...otherProps} size='small' color={color as MuiIconButtonProps['color']}>
       {icon}
     </IconButton>
+  );
+}
+
+export function CopyButton<T = any, S extends StrictRJSFSchema = RJSFSchema, F extends FormContextType = any>(
+  props: IconButtonProps<T, S, F>
+) {
+  const {
+    registry: { translateString },
+  } = props;
+  return (
+    <MuiIconButton
+      title={translateString(TranslatableString.CopyButton)}
+      {...props}
+      icon={<CopyIcon fontSize='small' />}
+    />
   );
 }