+++ packages/semantic-ui/src/TextareaWidget/TextareaWidget.tsx
@@ -34,7 +34,7 @@ export default function TextareaWidget<
   const semanticProps = getSemanticProps<T, S, F>({
     formContext,
     options,
-    defaultSchemaProps: { inverted: false },
+    defaultSchemaProps: { inverted: 'false' },
   });
   const { schemaUtils } = registry;
   // eslint-disable-next-line no-shadow